authorization implemented

+++ src/api/notes/handler.js
@@ -1,5 +1,3 @@
-const ClientError = require('../../exceptions/ClientError');
-
 class NotesHandler {
   constructor(service, validator) {
     this._service = service;
@@ -15,8 +13,11 @@ class NotesHandler {
   async postNoteHandler(request, h) {
     this._validator.validateNotePayload(request.payload);
     const { title = 'untitled', body, tags } = request.payload;
+    const { id: credentialId } = request.auth.credentials;
 
-    const noteId = await this._service.addNote({ title, body, tags });
+    const noteId = await this._service.addNote({
+        title, body, tags, owner: credentialId,
+      });
 
     const response = h.response({
       status: 'success',
@@ -29,8 +30,9 @@ class NotesHandler {
     return response;
   }
 
-  async getNotesHandler() {
-    const notes = await this._service.getNotes();
+  async getNotesHandler(request) {
+    const { id: credentialId } = request.auth.credentials;
+    const notes = await this._service.getNotes(credentialId);
     return {
       status: 'success',
       data: {
@@ -39,14 +41,12 @@ class NotesHandler {
     };
   }
 
-  async getNoteByIdHandler(request, h) {
+ async getNoteByIdHandler(request) {
     const { id } = request.params;
+    const { id: credentialId } = request.auth.credentials;
+
+    await this._service.verifyNoteOwner(id, credentialId);
     const note = await this._service.getNoteById(id);
-
-    if (!note) {
-      throw new ClientError('Catatan tidak ditemukan', 404);
-    }
-
     return {
       status: 'success',
       data: {
@@ -55,11 +55,12 @@ class NotesHandler {
     };
   }
 
-  async putNoteByIdHandler(request, h) {
+  async putNoteByIdHandler(request) {
     this._validator.validateNotePayload(request.payload);
-    const { id } = request.params;
+    const { id } = request.params; const { id: credentialId } = request.auth.credentials;
 
-    await this._service.editNoteById(id, request.payload);
+    await this._service.verifyNoteOwner(id, credentialId);
+    this._service.editNoteById(id, request.payload);
 
     return {
       status: 'success',
@@ -67,9 +68,12 @@ class NotesHandler {
     };
   }
 
-  async deleteNoteByIdHandler(request, h) {
+  async deleteNoteByIdHandler(request) {
     const { id } = request.params;
-    await this._service.deleteNoteById(id);
+    const { id: credentialId } = request.auth.credentials;
+
+    await this._service.verifyNoteOwner(id, credentialId);
+    this._service.deleteNoteById(id);
 
     return {
       status: 'success',
